requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.23.0
python-dotenv>=1.0.1
requests>=2.32.0
aiofiles>=23.2.1
Jinja2>=3.1.4
assemblyai>=0.33.0
google-generativeai>=0.7.0
gtts>=2.5.1
pydantic>=2.7.4
python-multipart>=0.0.9
SQLAlchemy>=2.0.31
